refacto: create an euro filter
