+++ gsl_programmation/tests/services/test_simulation_services.py
@@ -8,7 +8,7 @@ from gsl_core.tests.factories import (
     PerimetreRegionalFactory,
 )
 from gsl_programmation.models import Enveloppe, Simulation
-from gsl_programmation.services import SimulationService
+from gsl_programmation.service.simulation_service import SimulationService
 from gsl_programmation.tests.factories import (
     DetrEnveloppeFactory,
     DsilEnveloppeFactory,
